% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{gettree.hclust}
\alias{gettree.hclust}
\title{gettree.hclust}
\usage{
gettree.hclust(hc, root_index = NULL)
}
\arguments{
\item{hc}{result of hclust function}

\item{root_index}{the index of the node to start as root, generally not necessary to specify if one wants to get the full tree from hclust}
}
\description{
This function takes a hclust result and returns a tree structure
}

