% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gettree}
\alias{gettree}
\title{gettree}
\usage{
gettree(merge_data, root_index)
}
\arguments{
\item{merge_data}{'merge' object resulted from hclust function}

\item{root_index}{the largest index of the node in hclust tree}
}
\description{
gettree
}

