% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{heatmap.adjacency}
\alias{heatmap.adjacency}
\title{Generate heatmap from an adjacency matrix (symmetric)}
\usage{
heatmap.adjacency(m, zeroBasedIndex = TRUE, nodeLabels = NULL)
}
\description{
Return 2 json files for visualizing a heatmap: matrix and row dendrogram
}

