% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{heatmap.generic}
\alias{heatmap.generic}
\title{Generate heatmap from a generic mxn matrix}
\usage{
heatmap.generic(m, rowNodeType = "row", colNodeType = "col")
}
\description{
Generate heatmap from a generic mxn matrix
}

