% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{matrix2json}
\alias{matrix2json}
\title{matrix2json}
\usage{
matrix2json(m, mode = "undirected", weighted = TRUE,
  zeroBasedIndex = TRUE, node.attributes = NA)
}
\description{
Convert an adjacency matrix to json-compatible object
}

