\name{RVisLib-package}
\alias{RVisLib-package}
\alias{RVisLib}
\docType{package}
\title{
\packageTitle{RVisLib}
}
\description{
\packageDescription{RVisLib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RVisLib}
\packageIndices{RVisLib}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RVisLib}

Maintainer: \packageMaintainer{RVisLib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
