% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{shapes_rvis}
\alias{shapes_rvis}
\title{Scales for ggplot2 shapes}
\usage{
shapes_rvis(...)
}
\arguments{
\item{...}{common discrete scale parameters: name, breaks, labels, na.value, limits and guide. See discrete_scale for more details}
}
\description{
shapes_rvis maps discrete variables to shapes. The major difference between ggplot2 scale_shape function and shapes_rvis is the number of shapes. With this function you can map up to 17 shapes. If you have more than 17 levels, they will not be displayed - if you have more than 17 levels, it is probably not the best approach to use shapes for mapping.
}
\examples{
# create a dataframe with 17 distinct values
d=data.frame(a=letters[1:17])
library(ggplot2)
# plot shapes
ggplot() +
  scale_x_discrete(name="") +
  scale_y_continuous(limits=c(0,1)) +
  geom_point(data=d, mapping=aes(x=a, y=0.5, shape=a), size=5)+
  shapes_rvis()

}
\seealso{
\code{\link{scale_shape}}, \code{\link{discrete_scale}}
}
