% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_rvis}
\alias{theme_rvis}
\title{Custom ggplot2 theme}
\usage{
theme_rvis(
  base_size = 8,
  base_family = "Helvetica",
  x.text.angle = 0,
  spacing = 0.1,
  legend.pos = "top",
  legend.dir = NA,
  leg.box.order = NA,
  just = NA,
  just.par = NA
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{x.text.angle}{rotation angle of x axis tick labels. Use 90 for vertical orientation. Defaults to 0.}

\item{spacing}{space between facets (in cm). Defaults to 0.1.}

\item{legend.pos}{character specifying the position of legend. 'top','bottom','left','right', or a vector of two. Defaults to 'top'.}

\item{legend.dir}{direction of the keys in a legend.}

\item{leg.box.order}{direction in which multiple legends are displayed}

\item{just}{justification parameter as accepted by legend.box.just}

\item{just.par}{justification parameter as accepted by legend.justification}
}
\description{
Custom ggplot2 theme
}
\examples{
library(ggplot2)
# Plot with the default theme
ggplot(mpg, aes(x = class, fill = drv)) + geom_bar() +
  ggtitle('Plot with the default theme')
# Plot with theme_rvis
ggplot(mpg, aes(x = class, fill = drv)) + geom_bar() +
  theme_rvis(x.text.angle = 60, legend.pos = "top")+
  ggtitle('Using custom theme function')

}
\seealso{
\code{\link{theme}}
}
