% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tsv2jsoncy}
\alias{tsv2jsoncy}
\title{tsv2jsoncy}
\usage{
tsv2jsoncy(tsv_file, mode = "signed", zeroBasedIndex = TRUE)
}
\description{
Convert a tsv-formatted directed network to json-compatible object that could be imported directly by cytoscape
}

