% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{heatmap.adjacency}
\alias{heatmap.adjacency}
\title{Generate heatmap from an adjacency matrix (symmetric)}
\usage{
heatmap.adjacency(m)
}
\description{
Return 3 json files for visualizing a heatmap: matrix, row dendrogram and column dendrogram
}

